[1.1.X] Fixed #15284 - improved example jQuery code for adding X-CSRF-Token
Using the ajaxSend event is better than beforeSend, because the beforeSend
callback can have only one value, which makes it painful if it is needed by
multiple bits of javascript.

Thanks to LukeMaurer for report and initial patch.

Backport of [15515] from trunk.

This is backported to 1.1.X because it really belongs with security patch [15466]

git-svn-id: http://code.djangoproject.com/svn/django/branches/releases/1.1.X@15518 bcc190cf-cafb-0310-a4f2-bffc1f526a37